Auth

+++ src/todos/todos.service.ts
@@ -1,9 +1,10 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateTodoInput } from './dto/create-todo.input';
-import { UpdateTodoInput } from './dto/update-todo.input';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Todo } from './entities/todo.entity';
+import { UpdateTodoInput } from './dto/update-todo.input';
+import { User } from '../users/entities/user.entity';
 
 @Injectable()
 export class TodosService {
@@ -12,31 +13,42 @@ export class TodosService {
     private todoRepository: Repository<Todo>,
   ) {}
 
-  create(createTodoInput: CreateTodoInput): Promise<Todo> {
-    const todo = this.todoRepository.create(createTodoInput);
+  create(createTodoInput: CreateTodoInput, user: User): Promise<Todo> {
+    const todo = this.todoRepository.create({
+      ...createTodoInput,
+      user,
+      userId: user.id,
+    });
     return this.todoRepository.save(todo);
   }
 
-  findAll(): Promise<Todo[]> {
-    return this.todoRepository.find();
+  findAll(user: User): Promise<Todo[]> {
+    return this.todoRepository.find({ where: { userId: user.id } });
   }
 
-  async findOne(id: number): Promise<Todo> {
-    const todo = await this.todoRepository.findOne({ where: { id } });
+  async findOne(id: number, user: User): Promise<Todo> {
+    const todo = await this.todoRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!todo) {
       throw new NotFoundException(`Todo with ID ${id} not found`);
     }
     return todo;
   }
 
-  async update(id: number, updateTodoInput: UpdateTodoInput): Promise<Todo> {
-    await this.todoRepository.update(id, updateTodoInput);
-    return this.findOne(id);
+  async update(
+    id: number,
+    updateTodoInput: UpdateTodoInput,
+    user: User,
+  ): Promise<Todo> {
+    await this.findOne(id, user);
+    await this.todoRepository.update({ id, userId: user.id }, updateTodoInput);
+    return this.findOne(id, user);
   }
 
-  async remove(id: number): Promise<Todo> {
-    const todo = await this.findOne(id);
-    await this.todoRepository.delete(id);
+  async remove(id: number, user: User): Promise<Todo> {
+    const todo = await this.findOne(id, user);
+    await this.todoRepository.delete({ id, userId: user.id });
     return todo;
   }
 }
